feat: add iOS platform support (Phase 1 - simulator)
- Add iOS platform detection and framework linking in CMakeLists.txt
- Create platform/ios/ with sokol/imgui impl, platform functions,
  and stubs for FileDialog/VideoGrabber
- Share VideoPlayer, Sound, and FBO implementations with macOS (pure
  AVFoundation/AudioToolbox/Metal, no AppKit dependency)
- Add touch event handling (TOUCHES_BEGAN/MOVED/ENDED/CANCELLED) with
  mouse event mapping so existing mouse-based apps work on iOS
- Add metal_ios to shader language targets
- Add iOS build settings in trussc_app.cmake (Info-iOS.plist, bundle
  properties, code signing disabled for simulator)
- Compile sound sources as ObjC++ on iOS (miniaudio includes AVFoundation)

+++ trussc/include/tc/events/tcEventArgs.h
@@ -84,7 +84,7 @@ struct DragDropEventArgs {
 // Touch event arguments (for future use)
 // ---------------------------------------------------------------------------
 struct TouchEventArgs {
-    int id = 0;               // Touch ID
+    uintptr_t id = 0;         // Touch ID
     float x = 0.0f;
     float y = 0.0f;
     float pressure = 1.0f;
